make userLOcations a seperate class and add getters/setters
